Fix type inconsistencies after merge with main
Update parameter types for consistency:
- inBufferClear: int32_t -> size_t
- dmacChannelDisable: uint32_t -> uint8_t
- uartInterruptEnable: uint32_t -> uint8_t
- ecmConfigChannel/configChannelCT/configChannelV: int8_t -> int_fast8_t
- ds18b20StartSample: int32_t -> uint8_t

+++ src/configuration.c
@@ -75,7 +75,7 @@ static void     enterBootloader(void);
 static uint32_t getBoardRevision(void);
 static char    *getLastReset(void);
 static void     handleConfirmation(char c);
-static void     inBufferClear(int32_t n);
+static void     inBufferClear(size_t n);
 static size_t   inBufferTok(void);
 static void     printSettingCT(const int32_t ch);
 static void     printSettingDatalog(void);
