+++ src/driver_DMAC.c
@@ -28,7 +28,7 @@ volatile DmacDescriptor *dmacGetDescriptor(uint8_t ch) { return &dmacs[ch]; }
 
 void dmacCallbackBufferFill(void (*cb)(void)) { cbBufferFill = cb; }
 
-void dmacChannelDisable(uint32_t ch) {
+void dmacChannelDisable(uint8_t ch) {
   DMAC->CHID.reg = ch;
   DMAC->CHCTRLA.reg &= ~DMAC_CHCTRLA_ENABLE;
 }
